requirements.txt
streamlit>=1.28.0
openai-whisper>=20231117
pydub>=0.25.1
numpy>=1.24.0
ffmpeg-python>=0.2.0